scraper_api/pyproject.toml
[project]
name = "scraper-api"
version = "0.1.0"
description = "NovelBin scraper API for RTM"
authors = [
    {name = "RTM"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "cloudscraper (>=1.2.71,<2.0.0)",
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
